refactor(cli): extract security cluster into commands/security.ts (5.0-alpha)
First cut of the 5.0-alpha cli.ts god-module split. Moves the 10-handler
`security` command cluster (cmdSecurity + its 9 internally-routed
sub-handlers: prescan, prescan-diff, policy, scan-staged, scan-build,
scan-transcripts, costs, verify-pull, check-gitignore, clear-cache) out of
cli.ts into src/commands/security.ts, mirroring the existing
scan.ts/triage.ts/audit.ts pattern.

- cmdSecurity is the only export (the COMMANDS-table entry); the 9
  sub-handlers stay module-private, reached via cmdSecurity's process.argv[3]
  routing.
- cli.ts imports cmdSecurity back and the COMMANDS dispatch table stays the
  single source of truth — no change to the command surface.
- Prunes now-unused security-only imports from cli.ts (security-policy,
  bumblebee, scan-staged, scan-build, cost-monitor, post-pull-audit,
  findCommittedSensitive); keeps scanTranscripts/checkGitignore/patchGitignore
  (still used elsewhere).
- Self-contained lift: zero cli.ts-local helpers, zero cross-cluster cmd*
  refs; dynamic imports re-pathed ./ -> ../.

cli.ts: 6330 -> 5611 lines (-719). Typecheck, eslint (0 errors), full test
suite (2580), public-surface snapshot, and prettier all green.
